Changed Data Types in `HelpData`
Changed the data types of `numPangrams` and `numPerfectPangrams` to
`long` as that is what useful Stream opereations return.

+++ app/src/main/java/xterminators/spellingbee/model/HelpData.java
@@ -23,8 +23,8 @@ import org.apache.commons.lang3.tuple.Pair;
 public record HelpData(
     int numWords,
     int totalPoints,
-    int numPangrams,
-    int numPerfectPangrams,
+    long numPangrams,
+    long numPerfectPangrams,
     Map<Pair<Character, Integer>, Integer> startingLetterGrid,
     Map<String, Integer> startingLetterPairs
 ) {}
